QPID-3002: Added explicit casts required by some compiler versions
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@1070974 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/broker/ThresholdAlerts.cpp
@@ -128,9 +128,9 @@ void ThresholdAlerts::observe(Queue& queue, qpid::management::ManagementAgent& a
     //If no explicit threshold settings were given use 80% of any
     //limit from the policy.
     const QueuePolicy* policy = queue.getPolicy();
-    Option<uint32_t> countThreshold("qpid.alert_count", policy ? policy->getMaxCount()*0.8 : 0);
+    Option<uint32_t> countThreshold("qpid.alert_count", (uint32_t) (policy ? policy->getMaxCount()*0.8 : 0));
     countThreshold.addAlias("x-qpid-maximum-message-count");
-    Option<uint64_t> sizeThreshold("qpid.alert_size", policy ? policy->getMaxSize()*0.8 : 0);
+    Option<uint64_t> sizeThreshold("qpid.alert_size", (uint64_t) (policy ? policy->getMaxSize()*0.8 : 0));
     sizeThreshold.addAlias("x-qpid-maximum-message-size");
 
     observe(queue, agent, countThreshold.get(settings), sizeThreshold.get(settings), repeatInterval.get(settings));
